upgrade: flutter_map v8 and Dart 3.6 compatibility

+++ lib/src/splay/popup_spec_builder.dart
@@ -8,7 +8,7 @@ import 'package:supercluster/supercluster.dart';
 class PopupSpecBuilder {
   static PopupSpec forDisplacedMarker(
     DisplacedMarker displacedMarker,
-    int lowestZoom,
+    double lowestZoom,
   ) =>
       PopupSpec(
         namespace: SuperclusterLayer.popupNamespace,
